Enhance error log parsing and field definitions
Updated the regex pattern for better accuracy, adding `message`,
`first_stack_class`, `first_stack_msg`, and `caused_by` fields.

For better performance, the `log_body` (used by lnav to discover
fields) is limited to the first message of the log. The rest of the
log, if it is a stack trace, remains available in the full `log_text`.

Additionally, because the pull-replication plugin logs into the
`error_log` without a level, error messages from the plugin are
mapped to the 'error' level.

Change-Id: I3ab262b068037daabf62353174252381ad0b3a06